hotelflow/requirements.txt
asgiref==3.9.1
attrs==25.3.0
black==25.1.0
channels==4.3.1
channels_redis==4.3.0
click==8.2.1
Django==4.2.23
django-cors-headers==4.7.0
django-filter==25.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.28.0
drf-spectacular-sidecar==2025.8.1
inflection==0.5.1
isort==6.0.1
jsonschema==4.25.1
jsonschema-specifications==2025.4.1
msgpack==1.1.1
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
pillow==11.3.0
platformdirs==4.3.8
psycopg2-binary==2.9.10
PyJWT==2.10.1
python-decouple==3.8
PyYAML==6.0.2
redis==6.4.0
referencing==0.36.2
rpds-py==0.27.0
sqlparse==0.5.3
typing_extensions==4.14.1
uritemplate==4.2.0
